Create new env when conda exists but detected env doesn't

+++ scripts/remote_setup.sh
@@ -30,22 +30,35 @@ rm conda-envs.tar
 if [ -f "$CONDASH" ]; then
     echo ""
     echo "Conda installation found."
-    echo "source $CONDASH" >> .bashrc
+    #echo "source $CONDASH" >> .bashrc
     source $CONDASH
-    conda activate $ENV_NAME
-    conda env export > existing_env.yml
 
-    # We compare all the lines except the last one, because that one has the conda profiles,
-    # which will necessarily be different.
-    if cmp -s <(head -n -1 existing_env.yml) <(head -n -1 local_env.yml); then
-        echo "Imported env matches existing env. Skipping updates."
-        UPDATE_ENV=0
+    ENVS=$(conda env list | awk '{print $1}' )
+    if [[ $ENVS = *"$ENV_NAME"* ]]; then
+        # If env exists
+        echo "Env $ENV_NAME found."
+        conda activate $ENV_NAME
+        conda env export > existing_env.yml
+
+        # We compare all the lines except the last one, because that one has the conda profiles,
+        # which will necessarily be different.
+        if cmp -s <(head -n -1 existing_env.yml) <(head -n -1 local_env.yml); then
+            echo "Imported env matches existing env. Skipping updates."
+            UPDATE_ENV=0
+        else
+            echo "Imported env doesn't match existing env. Updating env..."
+            conda env update -q --file local_env.yml
+            UPDATE_ENV=1
+        fi
+        rm existing_env.yml
     else
-        echo "Imported env doesn't match existing env. Updating env..."
-        conda env update -q --file local_env.yml
+        # If env doesn't exit
+        echo "Env $ENV_NAME not found. Creating..."
+        conda env create -q --file local_env.yml
+        conda activate $ENV_NAME
         UPDATE_ENV=1
-    fi
-    rm existing_env.yml
+    fi;
+
 else
     echo "Conda installation not found. Installing..."
     wget -q https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
@@ -53,6 +66,7 @@ else
     rm Miniconda3-latest-Linux-x86_64.sh
     echo "source $CONDASH" >> .bashrc
     source $CONDASH
+    echo "Creating $ENV_NAME env..."
     conda env create -q --file local_env.yml
     conda activate $ENV_NAME
     UPDATE_ENV=1
@@ -73,13 +87,6 @@ $AWS s3 cp --quiet --recursive $S3_DATASETS_PATH $DATASETS_PATH
 printf "Importing results from $S3_RESULTS_PATH"
 $AWS s3 cp --quiet --recursive $S3_RESULTS_PATH $RESULTS_PATH
 
-# If the local and existing environments are different, update the env on the bucket
-if [ $UPDATE_ENV -eq 1 ]; then
-    echo ""
-    echo "Zipping conda env and saving to bucket..."
-    tar -cf /home/ubuntu/conda-envs.tar -C $CONDA_PATH .
-    $AWS s3 cp /home/ubuntu/conda-envs.tar s3://$BUCKET_NAME/
-fi
 
 cd repo
 echo ""
@@ -90,6 +97,14 @@ $@
 cd ..
 
 $AWS s3 sync $RESULTS_PATH $S3_RESULTS_PATH
+
+# If the local and existing environments are different, update the env on the bucket
+if [ $UPDATE_ENV -eq 1 ]; then
+    echo ""
+    echo "Zipping conda env and saving to bucket..."
+    tar -cf /home/ubuntu/conda-envs.tar -C $CONDA_PATH .
+    $AWS s3 cp /home/ubuntu/conda-envs.tar s3://$BUCKET_NAME/
+fi
 
 conda deactivate
 echo ""
